NFT card completed
